feat(models.py): fixed a bug

+++ fideo/models.py
@@ -12,6 +12,6 @@ class Share(models.Model):
     share_beta_factor = models.FloatField(("beta_factor"), default=0.00)
     share_historical = models.CharField(("historical"), max_length=255)
     share_compound = models.FloatField(("compound"), default=0.00)
-    share_market_cap = models.BigIntegerField(("market_cap") , default=0.00)
-    share_volume = models.BigIntegerField(("volume") , default=0.00)
+    share_market_cap = models.FloatField(("market_cap") , default=0.00)
+    share_volume = models.FloatField(("volume") , default=0.00)
     last_share_price = models.FloatField(("last_close_price"), default=0.00)
